pyproject.toml
[tool.poetry]
name = "mss-babelbert"
version = "0.1.0"
description = ""
authors = ["tommaso-green"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
wordfreq = "3.0.3"
torchmetrics = "0.7.2"
wandb = "0.12.11"
tokenizers = "0.10.3"
seaborn = "0.11.2"
scipy = "1.6.0"
scikit-learn = "1.0.2"
pytorch-metric-learning = "1.2.0"
plotly = "5.1.0"
pandas = "1.3.3"
omegaconf = "2.1.1"
numpy = "1.20.3"
matplotlib = "3.5.1"
hydra-core = "1.1.1"
GPUtil = "1.4.0"
datasets = "2.7.0"
adapter-transformers = "2.3.0"
mpire = "2.6.0"
pytorch-lightning = "1.7.0"
torch = "1.13.0"
lang2vec = "^1.1.2"
babelnet = {version = "1.0.1", extras = ["rpc"]}
nltk = "3.7"
tqdm = "^4.64.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
torch = { shell = "pip uninstall -y torch torchvision && pip install torch torchvision --extra-index-url https://download.pytorch.org/whl/cu116 && pip install torch_scatter -f https://data.pyg.org/whl/torch-1.13.0+cu116.html" }
